# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Plataforma de Auditoria Fiscal Preventiva (CFDI Intelligence)
  version: 1.0.0
  description: |
    API REST para la plataforma de deteccion automatica de riesgos fiscales.
    Permite gestionar empresas, ingestar CFDIs y estados de cuenta bancarios,
    conciliar movimientos, detectar riesgos fiscales, calcular scoring de
    cumplimiento y descargar CFDIs directamente del SAT via FIEL (e.firma).

    **Autenticacion**: Todos los endpoints (excepto /auth/register, /auth/login
    y /health) requieren un Bearer Token JWT obtenido al autenticarse.

    **Vencimiento del token**: 8 horas desde su emision.
servers:
  - url: https://api.cfdi-intelligence.com
    description: Produccion
  - url: http://localhost:8000
    description: Desarrollo local
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/empresas/{empresa_id}/emitidos:
    get:
      operationId: getEmitidos
      tags: [Emitidos]
      summary: CFDIs emitidos por la empresa en el periodo
      description: |
        Retorna los CFDIs emitidos por la empresa en el periodo especificado,
        clasificados segun la logica SAT:

        **Ingresos**:
        - ventas_servicios: CFDIs regulares de ingreso
        - anticipos: CFDIs con ClaveProdServ 84111506 + MetodoPago PUE + sin CFDI relacionado
        - facturas_con_anticipo: CFDIs de ingreso con TipoRelacion=07

        **Egresos**:
        - notas_credito: CFDIs de egreso regulares
        - aplicaciones_anticipo: CFDIs de egreso con FormaPago=30

        Incluye advertencias cuando una factura con anticipo no tiene su
        CFDI Egreso de aplicacion correspondiente.
      parameters:
        - $ref: '#/components/parameters/EmpresaId'
        - name: periodo
          in: query
          required: true
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
          description: Periodo fiscal YYYY-MM
          example: "2024-01"
      responses:
        '200':
          description: CFDIs emitidos clasificados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmitidosResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/empresas/{empresa_id}/recibidos:
    get:
      operationId: getRecibidos
      tags: [Emitidos]
      summary: CFDIs recibidos por la empresa en el periodo
      description: |
        Retorna los CFDIs recibidos (compras y gastos) por la empresa en el
        periodo especificado, con resumen de IVA acreditable.
      parameters:
        - $ref: '#/components/parameters/EmpresaId'
        - name: periodo
          in: query
          required: true
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
          description: Periodo fiscal YYYY-MM
          example: "2024-01"
      responses:
        '200':
          description: CFDIs recibidos clasificados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecibidosResponse'
              example:
                periodo: "2024-01"
                resumen:
                  subtotal: 72000.00
                  iva_acreditable: 11520.00
                  total: 83520.00
                  num_compras: 12
                  num_egresos: 1
                  vigentes: 12
                  canceladas: 1
                compras: []
                egresos: []
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  # ============================================================
  # DASHBOARD
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenido al registrarse o hacer login.
        Incluir en el header: Authorization: Bearer {token}
        Duracion: 8 horas.
  parameters:
    EmpresaId:
      name: empresa_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: ID unico de la empresa (UUID)
      example: "a1b2c3d4-0001-0001-0001-000000000001"
  responses:
    UnauthorizedError:
      description: Token JWT ausente, invalido o expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Se requiere autenticacion"
    ForbiddenError:
      description: El usuario no tiene acceso a este recurso
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Sin acceso a esta empresa"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Descripcion del error
...
    CFDIItem:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        serie_folio:
          type: string
          nullable: true
        fecha:
          type: string
          format: date
        rfc_receptor:
          type: string
          nullable: true
        nombre_receptor:
          type: string
          nullable: true
        subtotal:
          type: number
          format: float
        descuento:
          type: number
          format: float
        total:
          type: number
          format: float
        iva:
          type: number
          format: float
        metodo_pago:
          type: string
          nullable: true
          description: "PUE = Pago en una sola exhibicion, PPD = Pago en parcialidades"
        forma_pago:
          type: string
          nullable: true
        uso_cfdi:
          type: string
          nullable: true
        moneda:
          type: string
          nullable: true
        estado:
          type: string
          nullable: true
        estado_pago:
          type: string
          nullable: true
          enum: [pendiente, pagado_parcial, pagado_total]
        cfdi_relacionados:
          type: array
          items:
            type: object
        es_anticipo:
          type: boolean
        es_factura_con_anticipo:
          type: boolean
    EmitidosResponse:
      type: object
      properties:
        periodo:
          type: string
        empresa_rfc:
          type: string
        resumen:
          type: object
          properties:
            subtotal:
              type: number
              format: float
            iva_trasladado:
              type: number
              format: float
            total_facturado:
              type: number
              format: float
            vigentes:
              type: integer
            canceladas:
              type: integer
            num_tipo_i:
              type: integer
            num_tipo_e:
              type: integer
            num_tipo_p:
              type: integer
            total_cfdi_periodo:
              type: integer
            total_ingresos:
              type: number
              format: float
            total_facturas_con_anticipo:
              type: number
              format: float
            total_anticipos_acumulados:
              type: number
              format: float
            total_aplicaciones_anticipo:
              type: number
              format: float
            ingreso_neto_periodo:
              type: number
              format: float
            num_ingresos:
              type: integer
            num_anticipos:
              type: integer
            num_facturas_con_anticipo:
              type: integer
            num_egresos:
              type: integer
            advertencias:
              type: array
              items:
                type: object
                properties:
                  tipo:
                    type: string
                  uuid_factura:
                    type: string
                  mensaje:
                    type: string
        ingresos:
          type: object
          properties:
            ventas_servicios:
              type: array
              items:
                $ref: '#/components/schemas/CFDIItem'
            anticipos:
              type: array
              items:
                $ref: '#/components/schemas/CFDIItem'
            facturas_con_anticipo:
              type: array
              items:
                $ref: '#/components/schemas/CFDIItem'
        egresos:
          type: object
          properties:
            notas_credito:
              type: array
              items:
                $ref: '#/components/schemas/CFDIItem'
            aplicaciones_anticipo:
              type: array
              items:
                $ref: '#/components/schemas/CFDIItem'
    RecibidosResponse:
      type: object
      properties:
        periodo:
          type: string
        resumen:
          type: object
          properties:
            subtotal:
              type: number
              format: float
            iva_acreditable:
              type: number
              format: float
            total:
              type: number
              format: float
            num_compras:
              type: integer
            num_egresos:
              type: integer
            vigentes:
              type: integer
            canceladas:
              type: integer
        compras:
          type: array
          items:
            type: object
            properties:
              uuid:
                type: string
              serie_folio:
                type: string
                nullable: true
              fecha:
                type: string
                format: date
              rfc_emisor:
                type: string
              nombre_emisor:
                type: string
                nullable: true
              subtotal:
                type: number
                format: float
              total:
                type: number
                format: float
              iva:
                type: number
                format: float
              estado:
                type: string
        egresos:
          type: array
          items:
            type: object
...